An Inter-procedural Control Flow graph generation from a specific starting method.
In this CL, we are tring to generate a ICFG from "processCommandApdu" (which is the
central point for our payment related analysis).

Initially, we tried to figure out whether a specific method is present inside the
"Application Classes" (the classes present in the apk). Later, we tried to implement
a Breadth First Search (BFS) type algorithm and iterate over the methods that are
reachable from the starting method (which is in our case "processCommandApdu" method.
The reachable methods are found by checking each statement (Stmt) whether it is an
"InvokeStmt" or not from the body of each method of the queue (q).

The dot output is generated in "output.dot" file.

Current command to run: GRADLE_OPTS="-Xms1724m -Xmx8196m" ./gradlew run --args="AndroidCallGraph SPARK draw"